fix(ingest): snappy_compress size_t/uint64_t mismatch on Apple clang
bolt_parquet_write passed a `std::size_t*` where snappy_compress wants a
`uint64_t*`; on arm64 macOS these are distinct types (unsigned long vs
unsigned long long) and the build failed. Use uint64_t + a checked cast.

+++ src/ingest/bolt_parquet_write.cpp
@@ -514,11 +514,11 @@ bool maybe_compress(const std::uint8_t* src, std::size_t src_len,
     if (codec == 1u) {
         const std::size_t cap = snappy_max_compressed_len(src_len);
         dst->resize(cap);
-        std::size_t out_len = 0;
+        std::uint64_t out_len = 0;
         if (!snappy_compress(src, src_len, dst->data(), cap, &out_len)) {
             return false;
         }
-        dst->resize(out_len);
+        dst->resize(static_cast<std::size_t>(out_len));
         return true;
     }
     return false;
